Fix LUKS TPM
Added some logic for fixing tpm to unlock LUKs. Variable set to false so it's not attempted by default. I did this in case others actually use this script and don't want it attempted. So if you do use TPM to unlock LUKs and want it, please edit the variable towards the top of the script.